PR feedback increase code coverage
With the refactor the total code coverage dropped for the module. It was determined that dict of same data within testing classes was the issue. The data would cause
the coverage tool to count more lines that were uncovered. To resolve, either moved the data to a separate .json data file or collapsed prettified data to a single line.

This resulted in a drastic increase in coverage. A proper solution would be to have the coverage calculation ignore these lines and/or ignore test classes from the total coverage count.