Rework ECS translation to always produce a single query
Splitting the STIX query into multiple ECS queries when qualifiers are
encountered does not make sense with combining operators such as `AND` and `OR`.
Instead, we now process START/STOP qualifiers on-the-fly.

Also fix `AND` between observation expressions being converted to `OR`.

The output is still an array of queries, but always of length 1.